+++ net/asyncloop_impl.h
@@ -127,7 +127,7 @@ void AsyncLoop<SELECTOR, WORKER>::handleRead_(int const fd){
 	if (size <= 0)
 		return handleSocketOps_(fd, size);
 
-	buffer.push((size_t) size, inputBuffer_);
+	buffer.push(size, inputBuffer_);
 
 	if (buffer.size() > conf_maxPacketSize_)
 		return handleDisconnect_(fd, DisconnectStatus::ERROR);
@@ -169,7 +169,7 @@ void AsyncLoop<SELECTOR, WORKER>::handleWrite_(int const fd){
 	if (size <= 0)
 		return handleSocketOps_(fd, size);
 
-	buffer.pop((size_t) size);
+	buffer.pop(size);
 
 	buffer.restartTimer();
